requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
requests==2.31.0
scikit-learn==1.3.0
numpy==1.24.3
joblib==1.3.2
opencv-python==4.8.1.78
Pillow==10.0.1
tensorflow==2.13.0
keras==2.13.1
